Refactor out most conditionals
This should reduce the amount of unnecessary testing being done by
returning early when we know we can't handle further cases.

Of course, this comes with the overhead of a call to keyUnknown, so
maybe it's not actually performant.  So if truth be told, I just prefer
to not see so many conditionals, regardless of performance.